private vars and methods, simulation set builder, log_run() method for Simulation
