+++ .gitignore
@@ -1,2 +1,3 @@
 config.py
+simulation_logs
 __pycache__
